+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1497** : Fixed data list result paging.
+
+* Issue **#1492** : Fixed data list result paging.
+
 * Issue **#1513** : You can now view data in folders.
 
 * Issue **#1500** : Fixed data delete/restore behaviour.